CXP-1419: renamed and moved GetAllTestAppsQuery
